belimo-hackathon/requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3
scipy==1.11.2
